@@ -1,12 +1,12 @@
 name: qr_code_scanner
 description: QR code scanner that can be embedded inside flutter. It uses zxing in Android and MTBBarcode scanner in iOS.
-version: 0.5.2
+version: 0.7.0
 homepage: https://juliuscanute.com
 repository: https://github.com/juliuscanute/qr_code_scanner
 
 environment:
   sdk: '>=2.12.0 <3.0.0'
-  flutter: ">=1.10.0"
+  flutter: ">=1.12.0"
 
 dependencies:
   js: ^0.6.3
@@ -16,7 +16,7 @@ dependencies:
     sdk: flutter
 
 dev_dependencies:
-  pedantic: ^1.11.0
+  flutter_lints: ^1.0.4
 
 flutter:
   plugin:
