@@ -4,16 +4,16 @@ import 'barcode_format.dart';
 
 /// The [Barcode] object holds information about the barcode or qr code.
 ///
-/// [code] is the content of the barcode.
+/// [code] is the string-content of the barcode.
 /// [format] displays which type the code is.
-/// Only for Android, [rawBytes] gives a list of bytes of the result.
+/// Only for Android and iOS, [rawBytes] gives a list of bytes of the result.
 class Barcode {
   Barcode(this.code, this.format, this.rawBytes, this.byteArray);
 
-  final String code;
+  final String? code;
   final BarcodeFormat format;
 
-  /// Raw bytes are only supported by Android.
+  /// Raw bytes are only supported by Android and iOS.
   final List<int>? rawBytes;
   final Uint8List? byteArray;
 }
