@@ -56,7 +56,8 @@ public class QRView:NSObject,FlutterPlatformView {
                     self?.setDimensions(result,
                                         width: arguments["width"] ?? 0,
                                         height: arguments["height"] ?? 0,
-                                        scanArea: arguments["scanArea"] ?? 0,
+                                        scanAreaWidth: arguments["scanAreaWidth"] ?? 0,
+                                        scanAreaHeight: arguments["scanAreaHeight"] ?? 0,
                                         scanAreaOffset: arguments["scanAreaOffset"] ?? 0)
                 case "startScan":
                     self?.startScan(call.arguments as! Array<Int>, result)
@@ -84,7 +85,7 @@ public class QRView:NSObject,FlutterPlatformView {
         return previewView
     }
     
-    func setDimensions(_ result: @escaping FlutterResult, width: Double, height: Double, scanArea: Double, scanAreaOffset: Double) {
+    func setDimensions(_ result: @escaping FlutterResult, width: Double, height: Double, scanAreaWidth: Double, scanAreaHeight: Double, scanAreaOffset: Double) {
         // Then set the size of the preview area.
         previewView.frame = CGRect(x: 0, y: 0, width: width, height: height)
         
@@ -103,9 +104,9 @@ public class QRView:NSObject,FlutterPlatformView {
         }
 
         // Set scanArea if provided.
-        if (scanArea != 0) {
+        if (scanAreaWidth != 0 && scanAreaHeight != 0) {
             scanner?.didStartScanningBlock = {
-                self.scanner?.scanRect = CGRect(x: Double(midX) - (scanArea / 2), y: Double(midY) - (scanArea / 2), width: scanArea, height: scanArea)
+                self.scanner?.scanRect = CGRect(x: Double(midX) - (scanAreaWidth / 2), y: Double(midY) - (scanAreaHeight / 2), width: scanAreaWidth, height: scanAreaHeight)
 
                 // Set offset if provided.
                 if (scanAreaOffset != 0) {
@@ -117,6 +118,15 @@ public class QRView:NSObject,FlutterPlatformView {
         }
         return result(width)
         
+    }
+    
+    static func getBinaryData(descriptor: CIBarcodeDescriptor?) -> FlutterStandardTypedData? {
+        guard let descriptor = descriptor as? CIQRCodeDescriptor else { return nil }
+        let errorCorrectedPayload = descriptor.errorCorrectedPayload
+        var binary = Binary(data: errorCorrectedPayload)
+        let decoder = BinaryDecoder(symbolVersion: descriptor.symbolVersion)
+        decoder.decode(&binary)
+        return FlutterStandardTypedData(bytes: Data(bytes: decoder.bytes))
     }
     
     func startScan(_ arguments: Array<Int>, _ result: @escaping FlutterResult) {
@@ -151,7 +161,8 @@ public class QRView:NSObject,FlutterPlatformView {
                                         typeString = "EAN_8"
                                     case AVMetadataObject.ObjectType.ean13:
                                         typeString = "EAN_13"
-                                    case AVMetadataObject.ObjectType.itf14:
+                                    case AVMetadataObject.ObjectType.itf14,
+                                         AVMetadataObject.ObjectType.interleaved2of5:
                                         typeString = "ITF"
                                     case AVMetadataObject.ObjectType.pdf417:
                                         typeString = "PDF_417"
@@ -162,20 +173,40 @@ public class QRView:NSObject,FlutterPlatformView {
                                     default:
                                         return
                                 }
-                                var byteArray: [UInt8]? = nil
-                                if #available(iOS 11.0, *) {
-                                    if let descriptor = code.descriptor as? CIQRCodeDescriptor {
-                                        let errorCorrectedPayload = descriptor.errorCorrectedPayload
-                                        var binary = Binary(data: errorCorrectedPayload)
-                                        let symbolVersion = descriptor.symbolVersion
-                                        let decoder = BinaryDecoder(symbolVersion: symbolVersion)
-                                        decoder.decode(&binary)
-                                        byteArray = decoder.bytes
+                                let bytes = { () -> Data? in
+                                    if #available(iOS 11.0, *) {
+                                        switch (code.descriptor) {
+                                        case let qrDescriptor as CIQRCodeDescriptor:
+                                            return qrDescriptor.errorCorrectedPayload
+                                        case let aztecDescriptor as CIAztecCodeDescriptor:
+                                            return aztecDescriptor.errorCorrectedPayload
+                                        case let pdf417Descriptor as CIPDF417CodeDescriptor:
+                                            return pdf417Descriptor.errorCorrectedPayload
+                                        case let dataMatrixDescriptor as CIDataMatrixCodeDescriptor:
+                                            return dataMatrixDescriptor.errorCorrectedPayload
+                                        default:
+                                            return nil
+                                        }
+                                    } else {
+                                        return nil
                                     }
-                                }
-                                          
-                                guard let stringValue = code.stringValue else { continue }
-                                let result = ["code": stringValue, "type": typeString, "byteArray": byteArray as Any] as [String : Any]
+                                }()
+                                
+                                let byteArray = Self.getBinaryData(descriptor: code.descriptor)
+                                
+                                let result = { () -> [String : Any]? in
+                                    guard let stringValue = code.stringValue else {
+                                        guard let safeBytes = bytes else {
+                                            return nil
+                                        }
+                                        return ["type": typeString, "rawBytes": safeBytes, "byteArray": byteArray]
+                                    }
+                                    guard let safeBytes = bytes else {
+                                        return ["code": stringValue, "type": typeString, "byteArray": byteArray]
+                                    }
+                                    return ["code": stringValue, "type": typeString, "rawBytes": safeBytes, "byteArray": byteArray]
+                                }()
+                                guard result != nil else { continue }
                                 if allowedBarcodeTypes.count == 0 || allowedBarcodeTypes.contains(code.type) {
                                     self?.channel.invokeMethod("onRecognizeQR", arguments: result)
                                 }
@@ -188,9 +219,6 @@ public class QRView:NSObject,FlutterPlatformView {
                     let scanError = FlutterError(code: "unknown-error", message: "Unable to start scanning", details: error)
                     result(scanError)
                 }
-            } else {
-                let error = FlutterError(code: "cameraPermission", message: "Permission denied to access the camera", details: nil)
-                result(error)
             }
         })
     }
