@@ -102,10 +102,12 @@ class _QRViewState extends State<QRView> {
     return Stack(
       children: [
         _getPlatformQrView(),
-        Container(
+        Padding(
           padding: widget.overlayMargin,
-          decoration: ShapeDecoration(
-            shape: widget.overlay!,
+          child: Container(
+            decoration: ShapeDecoration(
+              shape: widget.overlay!,
+            ),
           ),
         )
       ],
@@ -127,7 +129,7 @@ class _QRViewState extends State<QRView> {
             onPlatformViewCreated: _onPlatformViewCreated,
             creationParams:
                 _QrCameraSettings(cameraFacing: widget.cameraFacing).toMap(),
-            creationParamsCodec: StandardMessageCodec(),
+            creationParamsCodec: const StandardMessageCodec(),
           );
           break;
         case TargetPlatform.iOS:
@@ -136,7 +138,7 @@ class _QRViewState extends State<QRView> {
             onPlatformViewCreated: _onPlatformViewCreated,
             creationParams:
                 _QrCameraSettings(cameraFacing: widget.cameraFacing).toMap(),
-            creationParamsCodec: StandardMessageCodec(),
+            creationParamsCodec: const StandardMessageCodec(),
           );
           break;
         default:
@@ -188,11 +190,11 @@ class QRViewController {
         case 'onRecognizeQR':
           if (call.arguments != null) {
             final args = call.arguments as Map;
-            final code = args['code'] as String;
+            final code = args['code'] as String?;
             final rawType = args['type'] as String;
             // Raw bytes are only supported by Android.
             final rawBytes = args['rawBytes'] as List<int>?;
-            final byteArray = Uint8List.fromList(List<int>.from(args['byteArray']));
+            final byteArray = args['byteArray'] as Uint8List?;
             final format = BarcodeTypesExtension.fromString(rawType);
             if (format != BarcodeFormat.unknown) {
               final barcode = Barcode(code, format, rawBytes, byteArray);
@@ -329,21 +331,44 @@ class QRViewController {
       {QrScannerOverlayShape? overlay}) async {
     if (defaultTargetPlatform == TargetPlatform.iOS) {
       // Add small delay to ensure the render box is loaded
-      await Future.delayed(Duration(milliseconds: 300));
+      await Future.delayed(const Duration(milliseconds: 300));
       if (key.currentContext == null) return false;
       final renderBox = key.currentContext!.findRenderObject() as RenderBox;
       try {
         await channel.invokeMethod('setDimensions', {
           'width': renderBox.size.width,
           'height': renderBox.size.height,
-          'scanArea': overlay?.cutOutSize ?? 0,
+          'scanAreaWidth': overlay?.cutOutWidth ?? 0,
+          'scanAreaHeight': overlay?.cutOutHeight ?? 0,
           'scanAreaOffset': overlay?.cutOutBottomOffset ?? 0
         });
         return true;
       } on PlatformException catch (e) {
         throw CameraException(e.code, e.message);
       }
+    } else if (defaultTargetPlatform == TargetPlatform.android) {
+      if (overlay == null) {
+        return false;
+      }
+      await channel.invokeMethod('changeScanArea', {
+        'scanAreaWidth': overlay.cutOutWidth,
+        'scanAreaHeight': overlay.cutOutHeight,
+        'cutOutBottomOffset': overlay.cutOutBottomOffset
+      });
+      return true;
     }
     return false;
+  }
+
+  //Starts/Stops invert scanning.
+  Future<void> scanInvert(bool isScanInvert) async {
+    if (defaultTargetPlatform == TargetPlatform.android) {
+      try {
+        await _channel
+            .invokeMethod('invertScan', {"isInvertScan": isScanInvert});
+      } on PlatformException catch (e) {
+        throw CameraException(e.code, e.message);
+      }
+    }
   }
 }
