@@ -1,5 +1,5 @@
-import 'package:flutter/widgets.dart';
+import 'package:flutter/material.dart';
 import 'package:qr_code_scanner/src/types/camera.dart';
 
 Widget createWebQrView({onPlatformViewCreated, CameraFacing? cameraFacing}) =>
-    throw UnimplementedError();
+    const SizedBox();
